backend/requirements.txt
asgiref==3.8.1
boto3==1.35.92
botocore==1.35.92
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
httpie==3.2.4
idna==3.10
inflection==0.5.1
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
Pygments==2.18.0
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
s3transfer==0.10.4
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.17.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
